DropIn Game mode refactoring
