--- src/controller/ui/ui/components/Notifiable.java
@@ -1,9 +1,0 @@
-package controller.ui.ui.components;
-
-/**
- * Created by rkessler on 2017-03-30.
- */
-public interface Notifiable {
-
-    void notifyObservers();
-}
