+++ src/controller/ui/ui/components/SequenceGameStates.java
@@ -1,0 +1,144 @@
+package controller.ui.ui.components;
+
+import common.TotalScaleLayout;
+import controller.action.ActionBoard;
+import controller.ui.ui.customized.Button;
+import controller.ui.ui.customized.ToggleButton;
+import data.Rules;
+import data.communication.GameControlData;
+import data.hl.HL;
+import data.states.AdvancedData;
+import data.values.SecondaryGameStates;
+
+import javax.swing.*;
+
+/**
+ * Created by rkessler on 2017-03-31.
+ */
+public class SequenceGameStates extends AbstractComponent {
+
+    protected static final String REFEREE_TIMEOUT = "Referee<br/>Timeout";
+    protected static final String FIRST_HALF = "First Half";
+    protected static final String SECOND_HALF = "Second Half";
+    protected static final String FIRST_HALF_SHORT = "1st Half";
+    protected static final String SECOND_HALF_SHORT = "2nd Half";
+    protected static final String FIRST_HALF_OVERTIME = "1st Extra";
+    protected static final String SECOND_HALF_OVERTIME = "2nd Extra";
+    protected static final String PENALTY_SHOOT = "Penalty Shots";
+    protected static final String PENALTY_SHOOT_SHORT = "Penalty";
+    protected static final String DROP_BALL = "Drop Ball";
+
+    protected JToggleButton firstHalf;
+    protected JToggleButton secondHalf;
+    protected JToggleButton firstHalfOvertime;
+    protected JToggleButton secondHalfOvertime;
+    protected JToggleButton penaltyShoot;
+    protected JToggleButton refereeTimeout;
+
+    // TODO - This needs to be moved to inheritence
+    protected JButton dropBall;
+
+
+    protected ButtonGroup halfGroup;
+
+    public SequenceGameStates() {
+        defineLayout();
+    }
+
+
+    public void defineLayout() {
+        if (!Rules.league.overtime) {
+            firstHalf = new ToggleButton(FIRST_HALF);
+            firstHalf.setSelected(true);
+            secondHalf = new ToggleButton(SECOND_HALF);
+            penaltyShoot = new ToggleButton(PENALTY_SHOOT);
+            refereeTimeout = new ToggleButton(REFEREE_TIMEOUT);
+
+            halfGroup = new ButtonGroup();
+            halfGroup.add(firstHalf);
+            halfGroup.add(secondHalf);
+            halfGroup.add(penaltyShoot);
+            halfGroup.add(refereeTimeout);
+
+        } else {
+            firstHalf = new ToggleButton(FIRST_HALF_SHORT);
+            firstHalf.setSelected(true);
+            secondHalf = new ToggleButton(SECOND_HALF_SHORT);
+            firstHalfOvertime = new ToggleButton(FIRST_HALF_OVERTIME);
+            secondHalfOvertime = new ToggleButton(SECOND_HALF_OVERTIME);
+            penaltyShoot = new ToggleButton(PENALTY_SHOOT_SHORT);
+            refereeTimeout = new ToggleButton(REFEREE_TIMEOUT);
+
+            halfGroup = new ButtonGroup();
+            halfGroup.add(firstHalf);
+            halfGroup.add(secondHalf);
+            halfGroup.add(firstHalfOvertime);
+            halfGroup.add(secondHalfOvertime);
+            halfGroup.add(penaltyShoot);
+            halfGroup.add(refereeTimeout);
+        }
+
+        firstHalf.addActionListener(ActionBoard.firstHalf);
+        secondHalf.addActionListener(ActionBoard.secondHalf);
+        if (Rules.league.overtime) {
+            firstHalfOvertime.addActionListener(ActionBoard.firstHalfOvertime);
+            secondHalfOvertime.addActionListener(ActionBoard.secondHalfOvertime);
+        }
+        penaltyShoot.addActionListener(ActionBoard.penaltyShoot);
+
+        if (Rules.league instanceof HL) {
+            dropBall = new Button(DROP_BALL);
+            dropBall.addActionListener(ActionBoard.dropBall);
+        }
+
+        TotalScaleLayout layout = new TotalScaleLayout(this);
+
+        layout.add(0.00, 0, 0.15, 1, firstHalf);
+        layout.add(0.15, 0, 0.15, 1, secondHalf);
+        layout.add(0.30, 0, 0.15, 1, firstHalfOvertime);
+        layout.add(0.45, 0, 0.15, 1, secondHalfOvertime);
+        layout.add(0.60, 0, 0.20, 1, penaltyShoot);
+        layout.add(0.80, 0, 0.20, 1, refereeTimeout);
+
+        this.setLayout(layout);
+
+    }
+
+    protected void updateDropBall(AdvancedData data) {
+        dropBall.setEnabled(ActionBoard.dropBall.isLegal(data));
+    }
+
+
+    @Override
+    @SuppressWarnings("Duplicates")
+    public void update(AdvancedData data) {
+
+        if (Rules.league instanceof HL) {
+            updateDropBall(data);
+        }
+        for (int i = 0; i < 2; i++) {
+//            name[i].setText(Teams.getNames(false)[data.team[i].teamNumber]);
+        }
+        firstHalf.setEnabled(ActionBoard.firstHalf.isLegal(data));
+        secondHalf.setEnabled(ActionBoard.secondHalf.isLegal(data));
+        if (Rules.league.overtime) {
+            firstHalfOvertime.setEnabled(ActionBoard.firstHalfOvertime.isLegal(data));
+            secondHalfOvertime.setEnabled(ActionBoard.secondHalfOvertime.isLegal(data));
+        }
+        penaltyShoot.setEnabled(ActionBoard.penaltyShoot.isLegal(data));
+        firstHalf.setSelected((data.secGameState == SecondaryGameStates.NORMAL)
+                && (data.firstHalf == GameControlData.C_TRUE));
+        secondHalf.setSelected((data.secGameState == SecondaryGameStates.NORMAL)
+                && (data.firstHalf != GameControlData.C_TRUE));
+        if (Rules.league.overtime) {
+            firstHalfOvertime.setSelected((data.secGameState == SecondaryGameStates.OVERTIME)
+                    && (data.firstHalf == GameControlData.C_TRUE));
+            secondHalfOvertime.setSelected((data.secGameState == SecondaryGameStates.OVERTIME)
+                    && (data.firstHalf != GameControlData.C_TRUE));
+        }
+        penaltyShoot.setSelected(data.secGameState == SecondaryGameStates.PENALTYSHOOT || data.previousSecGameState == SecondaryGameStates.PENALTYSHOOT);
+
+        refereeTimeout.setSelected(data.refereeTimeout);
+        refereeTimeout.setEnabled(ActionBoard.refereeTimeout.isLegal(data));
+    }
+}
