+++ src/data/states/GamePreparationData.java
@@ -2,6 +2,7 @@ package data.states;
 
 import data.Helper;
 import data.Rules;
+import data.hl.HLDropIn;
 import data.teams.TeamLoadInfo;
 import data.teams.TeamsLoader;
