+++ test/TestTeamLoader.java
@@ -63,6 +63,7 @@ public class TestTeamLoader {
         assertEquals(tli.identifier, 33);
         assertEquals(tli.colors.length, 2);
         assertEquals(tli.colors[0], "yellow");
+        // TODO could not resolve yellow
         assertEquals(tli.colors[1], "black");
         assertTrue(tli.icon != null);
     }
