+++ test_resources/config/hl_teen/teams.cfg
@@ -1,1 +1,1 @@
-33=NimbRo TeenSize,yellow,black
+33=NimbRo TeenSize
